move unit test to util
